// src/types/api.ts
export interface GithubComNezhahqNezhaModelValueArrayModelWAFApiMock {
    pagination: ModelPagination
    value: ModelWAFApiMock[]
}
// ...
export interface ModelPagination {
    limit: number
    offset: number
    total: number
}
// ...
export interface ModelWAFApiMock {
    block_identifier: number
    block_reason: number
    block_timestamp: number
    count: number
    ip: string
}